+++ src/Domain/Users/Services/AccountTransferService.cs
@@ -4,7 +4,7 @@ namespace Domain.Users.Services;
 
 public class AccountTransferService
 {
-    public void Transfer(Account from, Account to, int amount)
+    public void Transfer(Account from, Account to, long amount)
     {
         from.Debit(amount);
         to.Deposit(amount);
